requirements.txt
click==8.1.3
Flask==3.0.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.2
psycopg2-binary==2.9.5
Werkzeug==3.0.1
python-dotenv==1.0.1
sqlalchemy==2.0.27
pytest==6.2.4
